+++ src/models/ising_model.cpp
@@ -4,7 +4,7 @@
 // Constructor
 IsingModel::IsingModel(int n_sites, double temperature) {
     this->n_sites = n_sites;
-    J = std::vector<std::vector<double>>(n_sites, std::vector<double>(n_sites, 0.0));
-    H = std::vector<double>(n_sites, 0.0);
+    J = std::vector<std::vector<long double>>(n_sites, std::vector<long double>(n_sites, 0.0));
+    H = std::vector<long double>(n_sites, 0.0);
     this->temperature = temperature;
 }
